Update package to version 0.1.1 and add CHANGELOG
Updated the package and plugin version to 0.1.1 in package.json and plugin.json. This was done to reflect the new features and bug fixes introduced in this version. Also added a CHANGELOG for keeping track of all the past versions and the changes made in them. This includes addressing two GitHub issues, adding a feature and fixing a bug related to QMediaplayer backend.