Add new handlers for slicing canceled and backend error
The handlers for events such as slicing cancellation or a backend error have been added. These handlers make sure the audio stops playing in the cases of both events. This is done by calling the `_stopPlaying()` function, which has been newly added and encapsulates the process to stop and gradually fade out the audio.

Fixes #6